// src/restaurants/dto/update-restaurant.dto.ts
export class UpdateRestaurantInfoDto {
    id: number;
    uniqueId: string;
    restaurantName: string;
    image: string;
    address: string;
    pricePerson: string;
    website: string;
    businessHour: string;
    breakTime: string;
    restaurantPhone: string;
    country: string;
    countryAlias: string;
    state: string;
    stateAlias: string;
    status: string;
}